I2C DAC signal generator example
Created a program that can generate either a sine wave or a sawtooth
wave. The output frequency is adjustable from anywhere between
~0-600Hz. The program doesn’t use any delays, instead it dynamically
allocates an array with more values for a lower frequency with more
precision, or less values for a less accurate waveform at a higher
frequency. By using this method it maximizes speed and precision, and
only needs to call the sin function several times just to fill the
array at the start of the program (instead of calling it thousands of
times a second in the while loop that writes the values).